revert to alacritty as default sway term.
The Cosmic-term is a significant improvement in terms of BIDI
it shows Hebrew text properly even with pointing. Whereas
alacritty does not... it's backwards.

Problem with cosmic-term is there is a huge border/padding
that can't be configured as far as I can see. Also startup
time is significantly slower. So I'll use Alacritty for the
default term and add a sway shortcut to open up cosmic-term
where BIDI is needed.

Signed-off-by: David James McCorrie <djmccorrie@gmail.com>